Cleaning up compiler warnings

+++ src/monster2.c
@@ -2949,9 +2949,9 @@ void choose_new_monster(int m_idx, bool born, int r_idx)
         m_ptr->max_maxhp = (s16b)MIN(30000, hp);
     }
 
-    m_ptr->maxhp = (long)(m_ptr->maxhp * m_ptr->max_maxhp) / oldmaxhp;
+    m_ptr->maxhp = (int)m_ptr->maxhp * m_ptr->max_maxhp / oldmaxhp;
     if (m_ptr->maxhp < 1) m_ptr->maxhp = 1;
-    m_ptr->hp = (long)(m_ptr->hp * m_ptr->max_maxhp) / oldmaxhp;
+    m_ptr->hp = (int)m_ptr->hp * m_ptr->max_maxhp / oldmaxhp;
 }
